+++ tempodb/tempodb_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/grafana/tempo/pkg/tempopb"
 	"github.com/grafana/tempo/pkg/util/test"
 	"github.com/grafana/tempo/tempodb/backend/local"
+	"github.com/grafana/tempo/tempodb/encoding"
 	"github.com/grafana/tempo/tempodb/wal"
 	"github.com/stretchr/testify/assert"
 )
@@ -173,6 +174,113 @@ func TestRetention(t *testing.T) {
 	// retention again should clear it
 	r.(*readerWriter).doRetention()
 	checkBlocklists(t, blockID, 0, 0, rw)
+}
+
+func TestBlockCleanup(t *testing.T) {
+	tempDir, err := ioutil.TempDir("/tmp", "")
+	defer os.RemoveAll(tempDir)
+	assert.NoError(t, err, "unexpected error creating temp dir")
+
+	r, w, c, err := New(&Config{
+		Backend: "local",
+		Local: &local.Config{
+			Path: path.Join(tempDir, "traces"),
+		},
+		WAL: &wal.Config{
+			Filepath:        path.Join(tempDir, "wal"),
+			IndexDownsample: 17,
+			BloomFP:         .01,
+		},
+		BlocklistPoll: 0,
+	}, log.NewNopLogger())
+	assert.NoError(t, err)
+
+	c.EnableCompaction(&CompactorConfig{
+		ChunkSizeBytes:          10,
+		MaxCompactionRange:      time.Hour,
+		BlockRetention:          0,
+		CompactedBlockRetention: 0,
+	}, &mockSharder{})
+
+	blockID := uuid.New()
+
+	wal := w.WAL()
+	assert.NoError(t, err)
+
+	head, err := wal.NewBlock(blockID, testTenantID)
+	assert.NoError(t, err)
+
+	complete, err := head.Complete(wal, &mockSharder{})
+	assert.NoError(t, err)
+
+	err = w.WriteBlock(context.Background(), complete)
+	assert.NoError(t, err)
+
+	rw := r.(*readerWriter)
+
+	// poll
+	rw.pollBlocklist()
+
+	assert.Len(t, rw.blockLists[testTenantID], 1)
+
+	os.RemoveAll(tempDir + "/traces/" + testTenantID)
+
+	// poll
+	rw.pollBlocklist()
+
+	_, ok := rw.blockLists[testTenantID]
+	assert.False(t, ok)
+}
+
+func TestCleanMissingTenants(t *testing.T) {
+	tests := []struct {
+		name      string
+		tenants   []string
+		blocklist map[string][]*encoding.BlockMeta
+		expected  map[string][]*encoding.BlockMeta
+	}{
+		{
+			name:      "one missing tenant",
+			tenants:   []string{"foo"},
+			blocklist: map[string][]*encoding.BlockMeta{"foo": {{}}, "bar": {{}}},
+			expected:  map[string][]*encoding.BlockMeta{"foo": {{}}},
+		},
+		{
+			name:      "no missing tenants",
+			tenants:   []string{"foo", "bar"},
+			blocklist: map[string][]*encoding.BlockMeta{"foo": {{}}, "bar": {{}}},
+			expected:  map[string][]*encoding.BlockMeta{"foo": {{}}, "bar": {{}}},
+		},
+		{
+			name:      "all missing tenants",
+			tenants:   []string{},
+			blocklist: map[string][]*encoding.BlockMeta{"foo": {{}}, "bar": {{}}},
+			expected:  map[string][]*encoding.BlockMeta{},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			r, _, _, err := New(&Config{
+				Backend: "local",
+				Local: &local.Config{
+					Path: path.Join("/tmp", "traces"),
+				},
+				WAL: &wal.Config{
+					Filepath:        path.Join("/tmp", "wal"),
+					IndexDownsample: 17,
+					BloomFP:         .01,
+				},
+				BlocklistPoll: 0,
+			}, log.NewNopLogger())
+			assert.NoError(t, err)
+
+			rw := r.(*readerWriter)
+
+			rw.blockLists = tt.blocklist
+			rw.cleanMissingTenants(tt.tenants)
+			assert.Equal(t, rw.blockLists, tt.expected)
+		})
+	}
 }
 
 func checkBlocklists(t *testing.T, expectedID uuid.UUID, expectedB int, expectedCB int, rw *readerWriter) {